requirements.txt
dnspython>=2.7.0
openpyxl>=3.1.5
pandas>=2.2.3
streamlit>=1.45.1
xlrd>=2.0.1`